Add pubkey importing / seed generation (#172)
# Description
Add support for importing ed25519 keypairs and generating a new keypair
from seed bytes.

## Link to issue

#163 

## Type of change

- [x] New feature (non-breaking change that adds functionality)
- [x] This change requires a documentation update

## Test plan (required)

TODO

Demonstrate the code is solid. Which commands did you test with and what
are the expected results?
Which tests have you added or updated? Do the tests cover all of the
changes included in this PR?

## Screenshots/Screencaps

Please add previews of any UI Changes.