refactor: Add Gossipsub Message wrapper (#436)
# Description

This PR implements the following changes:

- [x] Add gosssipsub message wrapper
- [x] Rename `store_and_notify` to `publish_and_notify`

The message wrapper includes a header with a nonce to force the gossip
of duplicate receipts. We will likely expand on the header in future
work and make the nonce optional.

## Link to issue

Implements #421.

## Type of change

- [x] Refactor (non-breaking change that updates existing functionality)

## Test plan (required)

We've added a unit test to roundtrip a gossiped message to bytes and
back again. We also have a gossip notifications integration test to
confirm messages are still sent.

---------

Co-authored-by: Zeeshan Lakhani <zeeshan.lakhani@gmail.com>