chore: publish main package (#532)
# Description

Please include a summary of the change and which issue is fixed. Please
also include relevant motivation and context. List any dependencies that
are required for this change.

## Link to issue

Please add a link to any relevant issues/tickets.

## Type of change

- [ ] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Refactor (non-breaking change that updates existing functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [ ] Comments have been added/updated

Please delete options that are not relevant.

## Test plan (required)

Demonstrate the code is solid. Which commands did you test with and what
are the expected results?
Which tests have you added or updated? Do the tests cover all of the
changes included in this PR?

## Screenshots/Screencaps

Please add previews of any UI Changes.

---------

Co-authored-by: Zeeshan Lakhani <zeeshan.lakhani@gmail.com>